% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectedness.R
\docType{data}
\name{connectedness}
\alias{connectedness}
\title{US County Economic Connectednesss Data from Opportunity Insights}
\format{
A tibble with 3,089 observations and 3 variables:
\describe{
\item{location}{character variable for county and state name}
\item{population}{numeric variable for county population in 2018}
\item{connectedness}{numeric variable measuring economic connectedness}
}
}
\source{
\url{https://data.humdata.org/dataset/social-capital-atlas}
}
\usage{
connectedness
}
\description{
Data from \href{https://opportunityinsights.org/}{Opportunity Insights} about economic connectedness
across US counties. Connectedness is defined as two times the share of
high-SES friends among low-SES individuals, averaged over all low-SES
individuals in the county.
}
\author{
David Kane
}
\keyword{datasets}
