% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governors.R
\docType{data}
\name{governors}
\alias{governors}
\title{Election results and candidate longevity}
\format{
A tibble with 1,092 observations and 11 variables:
\describe{
\item{state}{character variable indicating the state in which an election took place}
\item{year}{integer variable indicating the year in which an election took place}
\item{first_name}{character variable indicating the first name of a candidate}
\item{last_name}{character variable indicating the last name of a candidate}
\item{party}{character variable indicating a candidate's party}
\item{sex}{character variable indicating a candidate's sex}
\item{died}{date variable indicating the candidate's date of death}
\item{status}{character variable indicating whether a candidate was the challenger or incumbent}
\item{win_margin}{double variable indicating the percentage margin by which the election was won
(positive values) or lost (negative values)}
\item{population}{number of people living in the state at the time of the election}
\item{death_age}{years a candidate lived}
\item{election_age}{years a candidate lived before the election took place}
\item{lived_after}{years a candidate lived after the election took place}
}
}
\source{
\url{https://doi.org/10.7910/DVN/IBKYRX}
}
\usage{
governors
}
\description{
This data is from the paper "Longevity Returns to Political
Office" by \href{https://doi.org/10.1017/psrm.2019.63}{Barfort, Klemmensen &
Larsen (2019)}. The purpose of this study was to explore how winning an
election influences the lifespan of a candidate. The authors collected data
on all U.S. gubernatorial elections from 1945 to 2012.
}
\details{
First, for a given election, only the two candidates who received the
highest number of votes were included. Second, candidates with unknown
dates of death were excluded, resulting in fewer observations for elections in
recent years, since most recent candidates are still alive. Third, in a few
instances, only the year of birth or death could be determined; in these cases,
the date was taken to be July 1 of that year.

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab governors \cr
   Number of rows \tab 1092 \cr
   Number of columns \tab 14 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 7 \cr
   Date \tab 1 \cr
   numeric \tab 6 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1 \tab 4 \tab 14 \tab 0 \tab 50 \tab 0 \cr
   first_name \tab 0 \tab 1 \tab 2 \tab 11 \tab 0 \tab 299 \tab 0 \cr
   last_name \tab 0 \tab 1 \tab 3 \tab 11 \tab 0 \tab 615 \tab 0 \cr
   party \tab 0 \tab 1 \tab 8 \tab 11 \tab 0 \tab 3 \tab 0 \cr
   sex \tab 0 \tab 1 \tab 4 \tab 6 \tab 0 \tab 2 \tab 0 \cr
   status \tab 0 \tab 1 \tab 9 \tab 10 \tab 0 \tab 2 \tab 0 \cr
   region \tab 0 \tab 1 \tab 4 \tab 9 \tab 0 \tab 4 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   died \tab 0 \tab 1 \tab 1946-12-21 \tab 2019-08-18 \tab 1996-07-20 \tab 683 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   year \tab 0 \tab 1 \tab 1964.85 \tab 13.38 \tab 1945.00 \tab 1954.00 \tab 1.962e+03 \tab 1.974e+03 \tab 2011.00 \tab ▇▆▃▂▁ \cr
   win_margin \tab 0 \tab 1 \tab 3.22 \tab 26.07 \tab -82.43 \tab -9.87 \tab 9.700e-01 \tab 1.218e+01 \tab 100.00 \tab ▁▃▇▁▁ \cr
   population \tab 0 \tab 1 \tab 3783555.08 \tab 4231731.71 \tab 145000.00 \tab 900000.00 \tab 2.427e+06 \tab 4.536e+06 \tab 27102238.00 \tab ▇▁▁▁▁ \cr
   election_age \tab 0 \tab 1 \tab 51.72 \tab 8.71 \tab 31.35 \tab 45.34 \tab 5.136e+01 \tab 5.748e+01 \tab 83.87 \tab ▂▇▆▂▁ \cr
   death_age \tab 0 \tab 1 \tab 79.95 \tab 11.16 \tab 41.63 \tab 73.38 \tab 8.190e+01 \tab 8.815e+01 \tab 101.72 \tab ▁▂▅▇▃ \cr
   lived_after \tab 0 \tab 1 \tab 28.23 \tab 13.38 \tab 0.13 \tab 17.57 \tab 2.960e+01 \tab 3.867e+01 \tab 60.42 \tab ▃▆▇▆▂ \cr
}
}
\author{
David Kane
}
\keyword{datasets}
