% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\docType{data}
\name{mail}
\alias{mail}
\title{Philadelphia experiment on mail-in voting}
\format{
A tibble with 935,707 observations and 10 variables:
\describe{
\item{treatment}{}
\item{voted}{}
\item{voted_mail}{}
\item{applied_mail}{}
\item{applied_date}{}
\item{party}{}
\item{age}{}
\item{sex}{}
\item{pred_white}{}
\item{pred_black}{}
}
}
\source{
\url{https://doi.org/10.7910/DVN/HUUEGI}
}
\usage{
mail
}
\description{

}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab mail \cr
   Number of rows \tab 935707 \cr
   Number of columns \tab 10 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   Date \tab 1 \cr
   character \tab 6 \cr
   factor \tab 1 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   applied_date \tab 785619 \tab 0.16 \tab 2020-03-10 \tab 2020-06-19 \tab 2020-05-18 \tab 53 \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   treatment \tab 0 \tab 1.00 \tab 4 \tab 12 \tab 0 \tab 3 \tab 0 \cr
   voted \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   voted_mail \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   applied_mail \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   party \tab 0 \tab 1.00 \tab 4 \tab 11 \tab 0 \tab 4 \tab 0 \cr
   sex \tab 239069 \tab 0.74 \tab 4 \tab 6 \tab 0 \tab 2 \tab 0 \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   age \tab 0 \tab 1 \tab FALSE \tab 5 \tab 50-: 217079, 18-: 208268, 30-: 206650, 65-: 161381 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pred_whi \tab 29318 \tab 0.97 \tab 0.32 \tab 0.39 \tab 0 \tab 0 \tab 0.06 \tab 0.73 \tab 1 \tab ▇▁▁▁▃ \cr
   pred_bla \tab 29318 \tab 0.97 \tab 0.45 \tab 0.44 \tab 0 \tab 0 \tab 0.32 \tab 0.95 \tab 1 \tab ▇▁▁▁▆ \cr
}
}
\author{
David Kane
}
\keyword{datasets}
