% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cces.R
\docType{data}
\name{cces}
\alias{cces}
\title{Data from the Cooperative Congressional Election Study}
\format{
A tibble with 470,755 observations and 12 variables:
\describe{
\item{year}{integer variable for year of survey}
\item{state}{character variable for state of residence for observation}
\item{gender}{charcater variable of "Female" and "Male"}
\item{age}{integer variable for age in years}
\item{education}{factor variable for level of education - takes values of "No HS",
"High School Graduate", "Some College", "2-Year", "4-Year", and "Post-Grad"}
\item{race}{character variable for racial identity}
\item{marstat}{character variable for marriage status}
\item{ideology}{factor variable for self-reported ideology - takes values of "Very Liberal",
"Liberal", "Moderate", "Conservative", "Very Conservative", and "Not Sure"}
\item{news}{character variable for level news/current events interest}
\item{econ}{character variable for retrospective report on the past year's economy}
\item{approval_ch}{character variable of approval of president - takes values of
"Strongly Approve", "Approve / Somewhat Approve", "Disapprove / Somewhat Disapprove",
"Never Heard / Not Sure", "Neither Approve Nor Disapprove"}
\item{approval}{numeric variable of approval for the current president on a 1-5 scale
with higher numbers indicating greater approval}
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/II2DB6}
}
\usage{
cces
}
\description{
Data for the approval ratings of voters to various government positions
combined with the demographic background of the voter. See
\href{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/II2DB6}{Kuriwaki
(2020)} for background and details. The raw code that was used to produce
the data is accessible at Kuriwaki's \href{https://github.com/kuriwaki/cces_cumulative}{Github.}
}
\details{
Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab cces \cr
   Number of rows \tab 470755 \cr
   Number of columns \tab 12 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 7 \cr
   factor \tab 2 \cr
   numeric \tab 3 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   state \tab 0 \tab 1.00 \tab 4 \tab 20 \tab 0 \tab 51 \tab 0 \cr
   gender \tab 0 \tab 1.00 \tab 4 \tab 6 \tab 0 \tab 2 \tab 0 \cr
   race \tab 0 \tab 1.00 \tab 5 \tab 15 \tab 0 \tab 8 \tab 0 \cr
   marstat \tab 1555 \tab 1.00 \tab 7 \tab 22 \tab 0 \tab 6 \tab 0 \cr
   news \tab 37007 \tab 0.92 \tab 10 \tab 17 \tab 0 \tab 5 \tab 0 \cr
   econ \tab 888 \tab 1.00 \tab 8 \tab 31 \tab 0 \tab 6 \tab 0 \cr
   approval_ch \tab 798 \tab 1.00 \tab 16 \tab 32 \tab 0 \tab 6 \tab 0 \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   ideology \tab 1774 \tab 1 \tab FALSE \tab 6 \tab Mod: 146842, Con: 108598, Lib: 82662, Ver: 54765 \cr
   education \tab 67 \tab 1 \tab FALSE \tab 6 \tab Hig: 130123, Som: 116433, 4-Y: 106727, Pos: 56623 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   year \tab 0 \tab 1 \tab 2012.92 \tab 3.84 \tab 2006 \tab 2010 \tab 2013 \tab 2016 \tab 2019 \tab ▅▅▅▇▆ \cr
   age \tab 0 \tab 1 \tab 49.60 \tab 16.47 \tab 18 \tab 36 \tab 51 \tab 62 \tab 109 \tab ▆▇▇▂▁ \cr
   approval \tab 798 \tab 1 \tab 2.68 \tab 1.67 \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab ▇▂▁▅▃ \cr
}
}
\author{
David Kane
}
\keyword{datasets}
