% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\docType{data}
\name{mail}
\alias{mail}
\title{Philadelphia experiment on mail-in voting}
\format{
A tibble with 935,707 observations and 10 variables:
\describe{
\item{treatment}{character variable indicating which of the 3 treatments was
employed before the 2020 primary election: 'No Postcard', 'Self', or 'Neighborhood'}
\item{voted}{character variable for whether respondent voted}
\item{voted_mail}{character variable for whether respondent voted by mail}
\item{applied_mail}{character variable for whether respondent applied for a mail ballot}
\item{applied_date}{date variable for the date that respondent's mail ballot application was received}
\item{party}{character variable for respondent's party registration}
\item{age}{factor variable for respondent's age group}
\item{sex}{character variable for respondent's sex}
\item{pred_white}{double variable for likelihood of respondent being white based on their name}
\item{pred_black}{double variable for likelihood of respondent being black based on their name}
}
}
\source{
\url{https://doi.org/10.7910/DVN/HUUEGI}
}
\usage{
mail
}
\description{
This data set comes from an experiment conducted in Philadelphia before the
2020 primary election. See also the 'Details' section below. The aim of the
study was to understand how facilitating mail ballot requests affects their
use. A total of over 935,000 registered voters were randomly assigned to
either a control group or one of two treatment groups. The results combine
information on voting behavior with voter's demographic background.
}
\details{
Pennsylvania's primary election took place on 2 June, 2020. Two weeks before
the election, on 18 May 2020, a group of randomly selected registrants in
Philadelphia received postcards with information about applying to vote by
mail. This was the first statewide federal election held since Pennsylvania
adopted universal access to mail ballots, so information about how to apply
was particularly likely to increase awareness of voting by mail.

Registrants were sent one of two postcards about mail ballots or no postcard.
The postcards conveyed information about the 26 May deadline to request a
mail ballot and included a message either encouraging voters to request a
ballot because “it’s safer for you to vote by mail!” or “it’s safer for
neighborhood to vote by mail!”. In all, 23,475 registered voters were
assigned to the “self” condition and 23,485 to the “neighborhood” condition,
with the remaining 888,785 in the control condition receiving no postcard.

 

Table: Data summary\tabular{ll}{
    \tab  \cr
   Name \tab mail \cr
   Number of rows \tab 935707 \cr
   Number of columns \tab 10 \cr
   _______________________ \tab  \cr
   Column type frequency: \tab  \cr
   character \tab 6 \cr
   Date \tab 1 \cr
   factor \tab 1 \cr
   numeric \tab 2 \cr
   ________________________ \tab  \cr
   Group variables \tab None \cr
}


\strong{Variable type: character}\tabular{lrrrrrrr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab empty \tab n_unique \tab whitespace \cr
   treatment \tab 0 \tab 1.00 \tab 4 \tab 12 \tab 0 \tab 3 \tab 0 \cr
   voted \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   voted_mail \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   applied_mail \tab 0 \tab 1.00 \tab 2 \tab 3 \tab 0 \tab 2 \tab 0 \cr
   party \tab 0 \tab 1.00 \tab 4 \tab 11 \tab 0 \tab 4 \tab 0 \cr
   sex \tab 239069 \tab 0.74 \tab 4 \tab 6 \tab 0 \tab 2 \tab 0 \cr
}


\strong{Variable type: Date}\tabular{lrrlllr}{
   skim_variable \tab n_missing \tab complete_rate \tab min \tab max \tab median \tab n_unique \cr
   applied_date \tab 785619 \tab 0.16 \tab 2020-03-10 \tab 2020-06-19 \tab 2020-05-18 \tab 53 \cr
}


\strong{Variable type: factor}\tabular{lrrlrl}{
   skim_variable \tab n_missing \tab complete_rate \tab ordered \tab n_unique \tab top_counts \cr
   age \tab 0 \tab 1 \tab FALSE \tab 5 \tab 50-: 217079, 18-: 208268, 30-: 206650, 65-: 161381 \cr
}


\strong{Variable type: numeric}\tabular{lrrrrrrrrrl}{
   skim_variable \tab n_missing \tab complete_rate \tab mean \tab sd \tab p0 \tab p25 \tab p50 \tab p75 \tab p100 \tab hist \cr
   pred_white \tab 29318 \tab 0.97 \tab 0.32 \tab 0.39 \tab 0 \tab 0 \tab 0.06 \tab 0.73 \tab 1 \tab ▇▁▁▁▃ \cr
   pred_black \tab 29318 \tab 0.97 \tab 0.45 \tab 0.44 \tab 0 \tab 0 \tab 0.32 \tab 0.95 \tab 1 \tab ▇▁▁▁▆ \cr
}
}
\author{
David Kane
}
\keyword{datasets}
